+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Arbitrary.hs
@@ -579,6 +579,7 @@ instance Arbitrary PoolVotingThresholds where
       <*> arbitrary
       <*> arbitrary
       <*> arbitrary
+      <*> arbitrary
 
 instance Arbitrary DRepVotingThresholds where
   arbitrary =