+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -66,6 +66,7 @@
 
 ### `testlib`
 
+* Add `integralToByteStringN`
 * Add `ByronKeyPair`
 * Add `Arbitrary` instances for Byron types: `Address`, `Attributes` and `AddrAttributes`
 * Add `Test.Cardano.Ledger.Core.JSON` with `roundTripJsonSpec`, `roundTripJsonEraSpec` and